# Faster JSON decoding of the bulk response