--- a/esutil/bulk_indexer.go
+++ b/esutil/bulk_indexer.go
@@ -20,7 +20,6 @@ package esutil
 import (
 	"bytes"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -33,6 +32,7 @@ import (
 
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/elastic/go-elasticsearch/v8/esapi"
+	"github.com/goccy/go-json"
 )
 
 // BulkIndexer represents a parallel, asynchronous, efficient indexer for Elasticsearch.