--- a/go.mod
+++ b/go.mod
@@ -2,4 +2,7 @@ module github.com/elastic/go-elasticsearch/v8
 
 go 1.13
 
-require github.com/elastic/elastic-transport-go/v8 v8.0.0-20211216131617-bbee439d559c
+require (
+	github.com/elastic/elastic-transport-go/v8 v8.0.0-20211216131617-bbee439d559c
+	github.com/goccy/go-json v0.10.0 // indirect
+)