@Anaethelion 2022-12-27
Thank you for raising this, however I'd prefer not to tie the whole library to anything more than the standard runtime.

Furthermore you can inject a custom decoder via the `BulkResponseJSONDecoder` interface in `BulkIndexerConfig.Decoder` which would allow you to achieve the same thing from your application code.